+++ CHANGELOG.rst
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 
 =====================
+22.03.08 - 22.03.08
+=====================
+* Added CSV for Block Volume Backups and Boot Volumes Backups
+* Added Volume Group Backup
+
+=====================
 22.02.22 - 22.02.22
 =====================
 * Added Support for ExaCS and ExaCC VMclusters in different compartment